+++ python/aistore/sdk/obj/obj_file/utils.py
@@ -85,8 +85,10 @@ def handle_chunked_encoding_error(
         ObjectFileReaderMaxResumeError: If the maximum number of resume attempts is exceeded.
     """
     resume_total = increment_resume(resume_total, max_resume, err)
+    obj_path = content_iterator.client.path
     logger.warning(
-        "Chunked encoding error (%s), retrying %d/%d",
+        "Chunked encoding error while reading '%s': (%s), retrying %d/%d",
+        obj_path,
         err,
         resume_total,
         max_resume,